refactor(Backend_v2): 重构任务执行逻辑并更新配置项
- 将任务执行逻辑从 yunji_api 模块移至 task_service 模块
- 更新任务执行流程，使用新的 task_service.run 方法
- 修改楼层作业柜名称，统一为"一层作业柜"、"二层作业柜"、"三层作业柜"
